19.3 Complete
Added webpack support. Added image optimization.
