+++ events.html
@@ -9,7 +9,7 @@
   <link rel="stylesheet" href="assets/css/bootstrap.css">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
   
-  <script src="./dist/main.bundle.js" defer ></script>
+  <script src="./dist/events.bundle.js" defer ></script>
 </head>
 <body>
   <nav class="navbar navbar-expand-lg navbar-light bg-light fixed-top">
