+++ index.html
@@ -10,7 +10,7 @@
   <link rel="stylesheet" href="assets/css/bootstrap.css">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
 
-  <script src="./dist/main.bundle.js" defer ></script>
+  <script src="./dist/app.bundle.js" defer ></script>
 
 </head>
 <body>
